+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 - Enhanced `Export-WsaJsonSpec` cmdlet to support deploying a Standard (single node) Workspace ONE Access via vRealize Suite Lifecycle Manager.
 - Enhanced `New-WSADeployment` cmdlet to support deploying a Standard (single node) Workspace ONE Access via vRealize Suite Lifecycle Manager.
-- Added `Enable-vRLIContentPack` cmdlet to support installation of content packs from the vRealize Log Insight in-product marketplace.
-- Added `Get-vRLIMarketplaceMetadata` cmdlet to support installation of content packs from the vRealize Log Insight in-product marketplace.
+- Added `Enable-vRLIContentPack` cmdlet to support install of content packs from the vRealize Log Insight in-product marketplace.
+- Added `Update-vRLIContentPack` cmdlet to support update of content packs from the vRealize Log Insight in-product marketplace.
+- Added `Get-vRLIMarketplaceMetadata` cmdlet to support install and upgrade of content packs from the vRealize Log Insight in-product marketplace.
 - Added `Install-vRLIContentPack` cmdlet to support installation of content packs from the vRealize Log Insight in-product marketplace.
-- Added `Get-vRLIContentPack` cmdlet to support installation of content packs from the vRealize Log Insight in-product marketplace.
+- Added `Get-vRLIContentPack` cmdlet to support install and upgrade of content packs from the vRealize Log Insight in-product marketplace.
 - Added `Get-vROpsLogForwardingConfig` to support verification of vRealize Operations log forwarding.
 - Added `Get-vROpsLogForwarding` to support verification of vRealize Operations log forwarding.
 - Updated Sample Scripts in the SampleScripts\ila folder, to support the VCF 4.5.x Planning and Prep Workbook: